refactor(fragments) gather all fragments in one package;
