+++ app/src/main/java/com/morihacky/android/rxjava/MainActivity.java
@@ -3,6 +3,8 @@ package com.morihacky.android.rxjava;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
+import com.morihacky.android.rxjava.fragments.MainFragment;
+import com.morihacky.android.rxjava.fragments.RotationPersistWorkerFragment;
 import com.morihacky.android.rxjava.rxbus.RxBus;
 
 public class MainActivity
