+++ app/src/main/java/com/morihacky/android/rxjava/fragments/DoubleBindingTextViewFragment.java
@@ -1,4 +1,4 @@
-package com.morihacky.android.rxjava;
+package com.morihacky.android.rxjava.fragments;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
