+++ app/src/main/java/com/morihacky/android/rxjava/rxbus/RxBusDemoFragment.java
@@ -6,7 +6,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import butterknife.ButterKnife;
-import com.morihacky.android.rxjava.BaseFragment;
+import com.morihacky.android.rxjava.fragments.BaseFragment;
 import com.morihacky.android.rxjava.R;
 
 public class RxBusDemoFragment
