+++ app/src/main/java/com/morihacky/android/rxjava/rxbus/RxBusDemo_TopFragment.java
@@ -7,7 +7,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
-import com.morihacky.android.rxjava.BaseFragment;
+import com.morihacky.android.rxjava.fragments.BaseFragment;
 import com.morihacky.android.rxjava.MainActivity;
 import com.morihacky.android.rxjava.R;
